backend/simulate_pipeline.py
#!/usr/bin/env python3
"""
Demonstration script showing how the application would work with valid credentials
"""
import os
import time
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def simulate_qdrant_connection():
    """Simulate connecting to Qdrant with credentials"""
    qdrant_url = os.getenv("QDRANT_URL")
    qdrant_api_key = os.getenv("QDRANT_API_KEY")
    cohere_api_key = os.getenv("COHERE_API_KEY")

    print("🔍 Checking environment variables...")
    print(f"  QDRANT_URL: {'✅ Configured' if qdrant_url and 'your-' not in qdrant_url else '❌ Using placeholder'}")
    print(f"  QDRANT_API_KEY: {'✅ Configured' if qdrant_api_key and 'your-' not in qdrant_api_key else '❌ Using placeholder'}")
    print(f"  COHERE_API_KEY: {'✅ Configured' if cohere_api_key and 'your-' not in cohere_api_key else '❌ Using placeholder'}")

    if 'your-' in (qdrant_url or '') or 'your-' in (qdrant_api_key or '') or 'your-' in (cohere_api_key or ''):
        print("\n⚠️  WARNING: You're using placeholder credentials!")
        print("   Please update your .env file with actual API keys and URLs.")
        print("   The application will simulate the process but won't connect to external services.")
        return False
    return True

def simulate_collection_creation():
    """Simulate creating a Qdrant collection"""
    print("\n🔧 Creating Qdrant collection...")
    print("   Collection name: robotic_book_embedding")
    print("   Vector size: 1024 dimensions")
    print("   Distance function: cosine")
    time.sleep(1)  # Simulate network delay
    print("   ✅ Collection created successfully!")

def simulate_embedding_generation(chunk_count):
    """Simulate generating embeddings for chunks"""
    print(f"\n🧮 Generating embeddings for {chunk_count} chunks...")
    print("   Using Cohere's embed-english-v3.0 model...")

    # Simulate processing with progress
    for i in range(0, min(10, chunk_count), max(1, chunk_count//10)):
        progress = min(100, (i + 1) * 100 // min(10, chunk_count))
        print(f"   Processed {i+1}/{min(10, chunk_count)} batches... {progress}%")
        time.sleep(0.1)  # Simulate processing time

    print(f"   ✅ Generated {chunk_count} embeddings successfully!")

def simulate_chunk_storage(chunk_count):
    """Simulate storing chunks in Qdrant"""
    print(f"\n📦 Storing {chunk_count} chunks in Qdrant...")

    # Simulate storing with progress
    for i in range(0, min(10, chunk_count), max(1, chunk_count//10)):
        progress = min(100, (i + 1) * 100 // min(10, chunk_count))
        print(f"   Stored {i+1}/{min(10, chunk_count)} batches... {progress}%")
        time.sleep(0.1)  # Simulate network delay

    print(f"   ✅ Stored {chunk_count} chunks successfully!")

def main():
    print("🤖 Humanoid Robotics RAG Pipeline - Simulation")
    print("=" * 50)

    # Check credentials
    has_valid_creds = simulate_qdrant_connection()

    # Load our analysis data
    total_chunks = 286  # From our previous analysis
    total_urls = 29

    print(f"\n📊 Data to be processed:")
    print(f"   Total URLs to process: {total_urls}")
    print(f"   Total chunks to create: {total_chunks}")
    print(f"   Average chunks per URL: {total_chunks/total_urls:.1f}")

    if has_valid_creds:
        print(f"\n🚀 Starting full pipeline with valid credentials...")

        # Simulate the main pipeline steps
        simulate_collection_creation()
        simulate_embedding_generation(total_chunks)
        simulate_chunk_storage(total_chunks)

        print(f"\n🏆 Pipeline completed successfully!")
        print(f"   - Collection: robotic_book_embedding")
        print(f"   - Total vectors stored: {total_chunks}")
        print(f"   - Total URLs processed: {total_urls}")

    else:
        print(f"\n🎯 Pipeline would process {total_urls} URLs and {total_chunks} chunks")
        print(f"   when valid credentials are provided.")
        print(f"   Each chunk would be stored with:")
        print(f"   - ID: calculated from URL index and chunk index")
        print(f"   - Vector: 1024-dimensional embedding")
        print(f"   - Payload: {{'url': source_url, 'text': chunk_content}}")

    print(f"\n💡 To run with actual credentials:")
    print(f"   1. Update your .env file with real API keys")
    print(f"   2. Run: source .venv/bin/activate && python src/main.py")

if __name__ == "__main__":
    main()