backend/pyproject.toml
[project]
name = "humanoid-robotics-rag"
version = "0.1.0"
description = "RAG pipeline for Humanoid Robotics Textbook - Extracts content, generates embeddings, and stores in Qdrant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.0.0",
    "qdrant-client>=1.7.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "lxml>=4.9.0",
    "trafilatura>=1.6.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
    "openai-agents>=0.1.0",
]

[project.scripts]
extract-content = "src.extraction.content_extractor:main"
generate-embeddings = "src.embeddings.embedding_generator:main"
store-vectors = "src.storage.vector_storage:main"
process-book = "src.main:main"
